+++ rust/processor/src/db/postgres/migrations/2025-01-21-033858_add_daily_base_volume_to_user_pools/up.sql
@@ -1,7 +1,7 @@
 -- Your SQL goes here
 DROP FUNCTION user_pools(provider text);
 CREATE FUNCTION user_pools(provider text) RETURNS TABLE(
-  transaction_version NUMERIC,
+  transaction_version BIGINT,
   sender VARCHAR(66),
   entry_function VARCHAR(200),
   transaction_timestamp TIMESTAMP,
